# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "holm",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/boxes/{id}/screens/{screen}/viewer/ticket": {
      "post": {
        "tags": [
          "viewer"
        ],
        "operationId": "ticket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "screen",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerTicket"
                }
              }
            }
          },
          "default": {
            "description": "The failure, as an ErrorBody",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "bad_request",
          "not_found",
          "gone",
          "denied",
          "screen_unavailable",
          "unsupported",
          "failed",
          "timeout",
          "unavailable",
          "transport",
          "internal"
        ]
      },
...
      "ViewerTicket": {
        "type": "object",
        "description": "Opens the viewer socket of one screen for a while, from a browser that can send no header.",
        "required": [
          "ticket",
          "expires_at_ms"
        ],
        "properties": {
          "control_socket": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "ticket": {
            "type": "string"
          },
          "view_socket": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}